-可以进行webDav的备份和恢复

+++ app/src/main/java/com/example/note/vm/NoteViewModel.kt
@@ -8,7 +8,12 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.note.bean.Note
 import com.example.note.db.NoteDao
+import com.example.note.util.fileToListT
 import com.example.note.util.writeToJson
+import com.hua.webdav.Authorization
+import com.hua.webdav.WebDav
+import com.hua.webdav.WebDavFile
+import com.hua.webdav.utils.getFileLength
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -25,6 +30,8 @@ class NoteViewModel @Inject constructor(
     private val dao: NoteDao,
     application: Application
 ) : AndroidViewModel(application) {
+
+    val authorization = Authorization("1297720454@qq.com", "a7awwp82sm6bis5f")
     init {
         searchDatabase("")
     }
@@ -47,6 +54,7 @@ class NoteViewModel @Inject constructor(
                     interPath.mkdir()
                 }
                 it.writeToJson("note.json", interPath.absolutePath)
+                if (it.isEmpty()) return@collectLatest
                 if (path.startsWith("content://")) {
                     val tree = DocumentFile.fromTreeUri(getApplication(), Uri.parse(path))!!
                     tree.findFile("note.json")?.delete()
@@ -59,8 +67,47 @@ class NoteViewModel @Inject constructor(
                     withContext(Dispatchers.Main) {
                         Toast.makeText(getApplication(), "备份成功", Toast.LENGTH_SHORT).show()
                     }
+                    val webDav = WebDav("https://dav.jianguoyun.com/dav/note", authorization)
+                    webDav.makeAsDir()
+                    val file = WebDavFile(
+                        "https://dav.jianguoyun.com/dav/note/note.json",
+                        authorization = authorization,
+                        displayName = "note.json",
+                        urlName = "https://dav.jianguoyun.com/dav/note/note.json",
+                        size = fileUri.getFileLength(getApplication()),
+                        contentType = "multipart/form-data",
+                        resourceType = "application/json",
+                        lastModify = System.currentTimeMillis()
+                    )
+                    if (file.exists()) file.delete()
+                    file.upload(getApplication(), fileUri, "multipart/form-data")
                 }
              }
+        }
+    }
+
+    fun restore(path: String) {
+        viewModelScope.launch {
+            val webDav = WebDav("https://dav.jianguoyun.com/dav/note/note.json", authorization)
+            if (!webDav.exists()) {
+                Toast.makeText(getApplication(), "备份文件不存在", Toast.LENGTH_SHORT).show()
+                return@launch
+            }
+            val tree = DocumentFile.fromTreeUri(getApplication(), Uri.parse(path))!!
+            tree.findFile("note.json")?.delete()
+            val fileUri = tree.createFile("application/json", "note.json")?.uri ?: return@launch
+            getApplication<Application>().contentResolver.openOutputStream(fileUri).use { ops ->
+                webDav.downloadInputStream().use { ins ->
+                    ins.copyTo(ops!!)
+                }
+            }
+            fileUri.fileToListT<Note>(getApplication())?.let {
+                dao.insertNotes(it)
+            }
+            withContext(Dispatchers.Main) {
+                Toast.makeText(getApplication(), "恢复成功", Toast.LENGTH_SHORT).show()
+            }
+            searchDatabase("")
         }
     }
